01_variable&add

+++ exercises/01_variable&add/main.cpp
@@ -6,8 +6,8 @@ int main(int argc, char **argv) {
     // TODO: 补全变量定义并打印加法运算
     // x ?
 
-    double x = std::stod(argv[1]);
-    // int x = std::stoi(argv[1]);
+    // double x = std::stod(argv[1]);
+    int x = std::stoi(argv[1]);
     std::cout << x << " + " << x << " = " << x + x << std::endl;    
     return 0;
 }
